be/requirements.txt
aiofile==3.9.0
aiohttp==3.11.9
aiosignal==1.3.1
amazon-transcribe==0.6.2
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
assemblyai==0.37.0
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
awscli==1.38.14
awscrt==0.16.26
azure-cognitiveservices-speech==1.42.0
boto3==1.37.14
botocore==1.37.14
cachetools==5.5.2
caio==0.9.21
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.1
dill==0.3.8
distlib==0.3.9
distro==1.9.0
Django==5.1.4
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docutils==0.16
fastapi==0.115.3
ffmpeg-python==0.2.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.0
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mysqlclient==2.2.6
networkx==3.4.2
numpy==2.1.2
oauthlib==3.2.2
openai==1.63.0
packaging==24.2
pipenv==2024.4.0
platformdirs==4.3.6
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.7.2
s3transfer==0.11.4
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.41.0
sympy==1.13.1
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.28.0
watchfiles==1.0.4
websockets==14.2
whitenoise==6.8.2
xxhash==3.5.0
yarl==1.18.3
zope.interface==7.2
